+++ src/lint.ts
@@ -80,6 +80,7 @@ export function analysisResult(diagnosticCollection: vscode.DiagnosticCollection
 export function Lint(diagnosticCollection: vscode.DiagnosticCollection, config: {[key:string]:any}, enableworkspace:boolean) {
     let filename = vscode.window.activeTextEditor.document.fileName
     let rootpath = vscode.workspace.rootPath
+    filename = filename.slice(rootpath.length + 1, filename.length);
     let cpplintOutput;
     if (enableworkspace) {
         cpplintOutput = runOnWorkspace(filename, rootpath, config);